@@ -16,7 +16,7 @@
     }
   ],
   "require": {
-    "php": "^8.2",
+    "php": "^8.3",
     "ext-curl": "*",
     "guzzlehttp/guzzle": "^7.0",
     "league/html-to-markdown": "^5.1",
